small balance refactor
