+++ pkg/balance/balance.go
@@ -1,0 +1,8 @@
+package balance
+
+type Balance struct {
+	CashUnlocked  uint64
+	CashLocked    uint64
+	TokenUnlocked uint64
+	TokenLocked   uint64
+}
